+++ app/src/main/java/com/truck/transfly/Model/RequestRating.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 class RequestRating {
 
     @SerializedName("rating")
-    var rating: Int? = 0;
+    var rating: Float? = 0f;
 
     @SerializedName("message")
     var message: String? = null;
